# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Employee Performance Rating API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /approvals:
    get:
      operationId: listApprovals
      tags: [approvals]
      security:
        - bearerAuth: []
      parameters:
        - name: ratedUserId
          in: query
          schema:
            type: string
        - name: teamId
          in: query
          schema:
            type: integer
        - name: quarter
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of approvals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Approval"
    post:
      operationId: createApproval
      tags: [approvals]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApprovalRequest"
      responses:
        "201":
          description: Approval created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
  /approvals/{approvalId}:
    patch:
      operationId: updateApproval
      tags: [approvals]
      security:
        - bearerAuth: []
      parameters:
        - name: approvalId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApprovalRequest"
      responses:
        "200":
          description: Approval updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
  /disputes:
    get:
      operationId: listDisputes
      tags: [disputes]
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Disputes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Approval"
  /disputes/{approvalId}:
    post:
      operationId: raiseDispute
      tags: [disputes]
      security:
        - bearerAuth: []
      parameters:
        - name: approvalId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeRequest"
      responses:
        "200":
          description: Dispute raised
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
    patch:
      operationId: resolveDispute
      tags: [disputes]
      security:
        - bearerAuth: []
      parameters:
        - name: approvalId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveDisputeRequest"
      responses:
        "200":
          description: Dispute resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [token, user]
    UserProfile:
      type: object
      properties:
        userId:
          type: string
        displayName:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [User, "Team Lead", Manager]
        level:
          type: string
          enum: [L1, L2, L3]
        teamId:
          type: integer
          nullable: true
        teamName:
          type: string
          nullable: true
      required: [userId, displayName, username, email, role, level]
...
    Approval:
      type: object
      properties:
        approvalId:
          type: integer
        itemId:
          type: integer
        itemName:
          type: string
          nullable: true
        teamId:
          type: integer
        ratedUserId:
          type: string
        ratedUserName:
          type: string
          nullable: true
        selfRatingValue:
          type: number
          nullable: true
        tlRatingValue:
          type: number
          nullable: true
        tlLgtmStatus:
          type: string
          enum: [Pending, Approved, Rejected]
        tlLgtmTimestamp:
          type: string
          format: date-time
          nullable: true
        tlLgtmByUserId:
          type: string
          nullable: true
        finalLgtmStatus:
          type: string
          enum: [Pending, Approved, Rejected]
        finalLgtmTimestamp:
          type: string
          format: date-time
          nullable: true
        finalLgtmByUserId:
          type: string
          nullable: true
        disputeStatus:
          type: boolean
        disputeComment:
          type: string
          nullable: true
        quarter:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        category:
          type: string
          nullable: true
        totalWeightedRating:
          type: number
          nullable: true
      required: [approvalId, itemId, teamId, ratedUserId, tlLgtmStatus, finalLgtmStatus, disputeStatus]
    CreateApprovalRequest:
      type: object
      properties:
        itemId:
          type: integer
        teamId:
          type: integer
        ratedUserId:
          type: string
        tlRatingValue:
          type: number
          minimum: 0.1
          maximum: 5
        quarter:
          type: string
          enum: [Q1, Q2, Q3, Q4]
        year:
          type: integer
      required: [itemId, teamId, ratedUserId, tlRatingValue, quarter, year]
    UpdateApprovalRequest:
      type: object
      properties:
        finalLgtmStatus:
          type: string
          enum: [Pending, Approved, Rejected]
        tlLgtmStatus:
          type: string
          enum: [Pending, Approved, Rejected]
    DisputeRequest:
      type: object
      properties:
        disputeComment:
          type: string
      required: [disputeComment]
    ResolveDisputeRequest:
      type: object
      properties:
        resolution:
          type: string
          enum: [Approved, Rejected]
        comment:
          type: string
      required: [resolution]
...