# lib/api-zod/src/generated/types/approval.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Employee Performance Rating API
 * OpenAPI spec version: 0.1.0
 */
import type { ApprovalFinalLgtmStatus } from "./approvalFinalLgtmStatus";
import type { ApprovalTlLgtmStatus } from "./approvalTlLgtmStatus";

export interface Approval {
  approvalId: number;
  itemId: number;
  itemName?: string | null;
  teamId: number;
  ratedUserId: string;
  ratedUserName?: string | null;
  selfRatingValue?: number | null;
  tlRatingValue?: number | null;
  tlLgtmStatus: ApprovalTlLgtmStatus;
  tlLgtmTimestamp?: Date | null;
  tlLgtmByUserId?: string | null;
  finalLgtmStatus: ApprovalFinalLgtmStatus;
  finalLgtmTimestamp?: Date | null;
  finalLgtmByUserId?: string | null;
  disputeStatus: boolean;
  disputeComment?: string | null;
  quarter?: string | null;
  year?: number | null;
  category?: string | null;
  totalWeightedRating?: number | null;
}