# lib/api-zod/src/generated/types/resolveDisputeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Employee Performance Rating API
 * OpenAPI spec version: 0.1.0
 */
import type { ResolveDisputeRequestResolution } from "./resolveDisputeRequestResolution";

export interface ResolveDisputeRequest {
  resolution: ResolveDisputeRequestResolution;
  comment?: string;
}